// src/Common/Hapi/Typings/search.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

import { FilterParametersInput } from "./Typyings";

// ====================================================
// GraphQL query operation: search
// ====================================================

export interface search_index_items_results_metadata {
  __typename: "Metadata";
  /**
   * Name of metadata field.
   */
  name: string | null;
  /**
   * Value of metadata field.
   */
  value: string | null;
}

export interface search_index_items_results_categories {
  __typename: "Categories";
  /**
   * Hierarchical list of categories names.
   */
  values: (string | null)[] | null;
}

export interface search_index_items_results {
  __typename: "ItemResult";
  /**
   * Id of the item. Autogenerated or explicitly set during insert.
   */
  id: string | null;
  /**
   * Internal id of item.
   */
  internalId: number;
  /**
   * Parent ids of item.
   */
  parentIds: (string | null)[] | null;
  /**
   * Parent level of item.
   */
  parentLevel: number;
  /**
   * Autogenerated Url of the item.
   */
  url: string | null;
  /**
   * Title of found item.
   */
  title: string | null;
  /**
   * Abstract of found item.
   */
  abstract: string | null;
  /**
   * Content of found item.
   */
  content: string | null;
  /**
   * Extracts of found item with marked hits.
   */
  extracts: string | null;
  /**
   * Revelance score.
   */
  relevance: number;
  /**
   * Insert date.
   */
  date: any;
  /**
   * Metadata fields describing item.
   */
  metadata: (search_index_items_results_metadata | null)[] | null;
  /**
   * List of categories associated with item.
   */
  categories: (search_index_items_results_categories | null)[] | null;
}

export interface search_index_items {
  __typename: "QueryResult";
  hits: number;
  results: (search_index_items_results | null)[] | null;
}

export interface search_index {
  __typename: "IndexQuery";
  items: search_index_items | null;
}

export interface search {
  index: search_index | null;
}

export interface searchVariables {
  indexId: number;
  filterParameters: FilterParametersInput;
}